requirements.txt
streamlit
streamlit-webrtc
streamlit-mic-recorder
google-genai
opencv-python-headless
Pillow
pydub
numpy